pyproject.toml
[project]
name = "tilores-sdk"
version = "0.1.0"
authors = [
  { name="Lukas Rieder", email="lukas@parlant.co" },
]
description = "This is the offical SDK to develop with the Tilores entity resolution database."
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["entity-resolution", "tilores", "graph-rag"]
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "requests>=2.32.3",
  "graphql-core>=3.2.3",
  "graphql-query>=1.4.0",
]
[project.optional-dependencies]
test = [
  "pytest==8.3.2",
]

[project.urls]
Homepage = "https://github.com/tilotech/python-tilores-sdk"
Issues = "https://github.com/tilotech/python-tilores-sdk/issues"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

